requirements.txt
Pillow
opencv-python
tqdm
rembg